Added new "BinaryBody" property to the "HttpRequest" class
The new property lets developers set the binary body of the HTTP request in situations where neither raw text or multipart data is sufficient.

The property is only added to the "HttpRequest" class for now, as adding it to the "IHttpRequest" interface would be a breaking change. We could add it to the interface later for v1.2.

Implements #4