Don't care about this for now
The compiler complains that the transitive dependency Newtonsoft.Json 9.0.1 has a high severity vulnerability. While this is true, Skybrud.Essentials and Skybrud.Essentials.Http will almost certainly be used in a context where a newer an unaffected version of Newtonsoft.Json is installed, so we don't care about this warning for now. For the upcoming v1.2.0 of Skybrud.Essentials, the dependency for Newtonsoft.Json will also be bumped to an unaffected version.